// src/app/page/master/desa-kelurahan/desa-kelurahan.interface.ts
import { EventEmitter } from '@angular/core';
export interface DesaKelurahan {

  kdKecamatan?: any;
  kdKotaKabupaten?: any;
  kdPropinsi?: any;
  kode?: any;
  kodeExternal?: string;
  kodePos?: string;
  namaDesaKelurahan?: string;
  namaExternal?: string;
  reportDisplay?: string;
  statusEnabled?: true;
  version?: any;
  noRec?: string;
  label?: any;
  value?: any;
  namaPropinsi?: string;
}